Move shared logic to DatabaseIndexApi class
