+++ packages/core/src/indexing/__tests__/build-indexing.test.ts
@@ -1,7 +1,6 @@
 import tmp from 'tmp-promise'
 import { buildIndexing, UnsupportedDatabaseProtocolError } from '../build-indexing.js'
-import { PostgresIndexApi } from '../postgres/postgres-index-api.js'
-import { SqliteIndexApi } from '../sqlite/sqlite-index-api.js'
+import { PostgresIndexApi, SqliteIndexApi } from '../database-index-api.js'
 import { LoggerProvider, Networks } from '@ceramicnetwork/common'
 import pgTest from '@databases/pg-test'
