+++ packages/core/src/indexing/__tests__/insertion-order.test.ts
@@ -1,7 +1,6 @@
 import knex, { Knex } from 'knex'
 import { StreamID } from '@ceramicnetwork/streamid'
-import { PostgresIndexApi } from '../postgres/postgres-index-api.js'
-import { SqliteIndexApi } from '../sqlite/sqlite-index-api.js'
+import { PostgresIndexApi, SqliteIndexApi } from '../database-index-api.js'
 import { readCsvFixture } from './read-csv-fixture.util.js'
 import { chunks } from '../../__tests__/chunks.util.js'
 import { InsertionOrder } from '../insertion-order.js'
