+++ packages/core/src/indexing/build-indexing.ts
@@ -1,6 +1,4 @@
-import type { DatabaseIndexApi } from './database-index-api.js'
-import { SqliteIndexApi } from './sqlite/sqlite-index-api.js'
-import { PostgresIndexApi } from './postgres/postgres-index-api.js'
+import { type DatabaseIndexApi, PostgresIndexApi, SqliteIndexApi } from './database-index-api.js'
 import knex from 'knex'
 import { DiagnosticsLogger, Networks } from '@ceramicnetwork/common'
 import * as fs from 'fs'
@@ -56,7 +54,10 @@ export function buildIndexing(
         // create dir if it doesn't exist
         // not strictly necessary here, but keeping it for backwards compatibility, as this directory
         // was created on startup before CDB-2008
-        fs.mkdirSync(connectionString.pathname.substring(0, connectionString.pathname.lastIndexOf('/')), { recursive: true })
+        fs.mkdirSync(
+          connectionString.pathname.substring(0, connectionString.pathname.lastIndexOf('/')),
+          { recursive: true }
+        )
       }
       const dbConnection = knex({
         client: 'sqlite3',
